(PC-13518)[PRO] refactor: migrate audience type into core/shared
