+++ pro/src/core/Offers/types.ts
@@ -43,11 +43,6 @@ export type Offer = {
   isShowcase?: boolean | null
 }
 
-export enum Audience {
-  INDIVIDUAL = 'individual',
-  COLLECTIVE = 'collective',
-}
-
 export type Option = {
   id: string
   displayName: string
