+++ pro/src/core/shared/types.ts
@@ -1,0 +1,4 @@
+export enum Audience {
+  INDIVIDUAL = 'individual',
+  COLLECTIVE = 'collective',
+}
