+++ pro/src/new_components/FilterTabs/FilterTabs.tsx
@@ -2,7 +2,7 @@ import cn from 'classnames'
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-import { Audience } from 'core/Offers/types'
+import { Audience } from 'core/shared/types'
 import { ReactComponent as LibraryIcon } from 'icons/library.svg'
 import { ReactComponent as UserIcon } from 'icons/user.svg'
