+++ pro/src/screens/Offers/NoResults/NoResults.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 
 import Icon from 'components/layout/Icon'
-import { Audience } from 'core/Offers/types'
+import { Audience } from 'core/shared/types'
 
 import styles from './NoResults.module.scss'
